requirements.txt
csv
cv2
numpy as np
os
paho.mqtt.publish as publish
RPi.GPIO as GPIO
time
picamera
base64
